+++ src/uk/co/tggl/pluckerpluck/multiinv/player/MIPlayer.java
@@ -114,7 +114,7 @@ public class MIPlayer {
             player.setHealth(MIYamlFiles.con.getHealth(player.getName(), group));
         } else {
             MIPlayerFile config = new MIPlayerFile(player, group);
-            int health = config.getHealth();
+            double health = config.getHealth();
             if(health < 0) {
                 health = 0;
             }
